+++ lib/beaker/host.rb
@@ -559,10 +559,7 @@ module Beaker
       end
 
       if opts.has_key?(:ignore) and not opts[:ignore].empty?
-        opts[:ignore].map! do |value|
-          "--exclude '#{value}'"
-        end
-        rsync_args << opts[:ignore].join(' ')
+        rsync_args << opts[:ignore].map { |value| "--exclude '#{value}'" }.join(' ')
       end
 
       # We assume that the *contents* of the directory 'from_path' needs to be