+++ CHANGELOG.md
@@ -26,5 +26,9 @@ git logs & PR history.
 
 - concept of `manual_test` and `manual_step`
 
+### Bug fixes
+
+- Fixed corruption of `opts[:ignore]` when using `rsync`
+
 # [3.25.0](https://github.com/puppetlabs/beaker/compare/3.24.0...3.25.0) - 2017-09-26